# docs/swagger.yaml
basePath: /api
definitions:
...
  internal_api.CreateWalletRequest:
    properties:
      currency:
        example: USD
        type: string
    required:
    - currency
    type: object
  internal_api.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
...
  internal_api.UpdateUserRequest:
    properties:
      display_name:
        example: John Doe
        type: string
      pin_required_for_transfer:
        example: true
        type: boolean
      profile_photo_url:
        example: https://example.com/photo.jpg
        type: string
    type: object
...
  verve_internal_models.User:
    description: User information including display name and profile settings
    properties:
      created_at:
        type: string
      display_name:
        example: John Doe
        type: string
      email:
        example: john.doe@example.com
        type: string
      id:
        example: 1
        type: integer
      pin_required_for_transfer:
        example: true
        type: boolean
      profile_photo_url:
        example: https://example.com/photo.jpg
        type: string
      provider:
        description: OAuth provider (google, okta, or local)
        example: google
        type: string
      provider_user_id:
        description: ID from the OAuth provider
        type: string
      roles:
        example:
        - '[''user'''
        - '''admin'']'
        items:
          type: string
        type: array
      updated_at:
        type: string
      username:
        example: john.doe
        type: string
    type: object
...
  verve_internal_models.Wallet:
    description: A digital wallet that can hold a specific currency
    properties:
      balance:
        example: 10000
        type: integer
      can_transfer:
        example: true
        type: boolean
      created_at:
        type: string
      currency:
        example: USD
        type: string
      deleted_at:
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      user_id:
        example: 1
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  title: Verve API
  version: "1.0"
paths:
...
  /user/{id}:
    get:
      description: Get details of a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/verve_internal_models.User'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user details
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the authenticated user's profile information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User details to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_api.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/verve_internal_models.User'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Forbidden - Can only update own profile
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user profile
      tags:
      - users
...
  /user/{id}/wallets:
    get:
      description: Get all wallets belonging to the authenticated user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/verve_internal_models.Wallet'
            type: array
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Forbidden - Can only view your own wallets
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user wallets
      tags:
      - wallets
    post:
      consumes:
      - application/json
      description: Create a new wallet for the authenticated user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Wallet details
        in: body
        name: wallet
        required: true
        schema:
          $ref: '#/definitions/internal_api.CreateWalletRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/verve_internal_models.Wallet'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Forbidden - Can only create wallets for yourself
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create wallet
      tags:
      - wallets
  /user/{user_id}/wallets/{wallet_id}:
    get:
      description: Get details of a specific wallet belonging to the authenticated
        user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Wallet ID
        in: path
        name: wallet_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/verve_internal_models.Wallet'
        "400":
          description: Invalid wallet ID
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Forbidden - Can only view your own wallet
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: Wallet not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get wallet details
      tags:
      - wallets
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"